# New_EcomAPI /swagger/swagger.yaml
swagger: '2.0'
info:
  title: 'CT Project E-Commerce API'
  description: 'An API to manage E-Commerce. This project was made by James Carlson as an assignment for Coding Temple.'
  version: '1.0.0'
host: 'ct-project-12-ecommerce-ci-cd.onrender.com'
schemes:
  - 'https'
consumes: 
  - 'application/json'
produces:
  - 'application/json'
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
security:
  - bearerAuth: []
...
paths: 
...
  /customers:
    post:
      tags: [
        "Customers"
      ]
      summary: "Create a new customer"
      description: "Will add a new customer to the database"
      parameters:
        - in: "body"
          name: "body"
          description: "Create customer object"
          required: true
          schema:
            $ref: "#/definitions/CustomerRequest"
      responses:
        201:
          description: 'Customer Created'
          schema:
            $ref: "#/definitions/CustomerResponse"
        400:
          description: 'Username Taken'
          examples:
            application/json:
              error: "Username is already taken. Please create a unique username."
    get:
      tags: [
        "Customers"
      ]
      summary: "Get Customers Endpoint"
      description: "Get all customer data from customer table"
      parameters:
        - in: query
          name: page
          type: "integer"
          description: "Page of customer objects"
          required: false
        - in: query
          name: per_page
          type: "integer"
          description: "Number of customer objects per page"
          required: false
      responses:
        200:
          description: "Successful Response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CustomerResponse"
...
  /orders:
    get:
      tags: [
        "Orders"
      ]
      summary: "Get Orders Endpoint"
      description: "Get all order data from order table"
      parameters:
        - in: query
          name: page
          type: "integer"
          description: "Page of order objects"
          required: false
        - in: query
          name: per_page
          type: "integer"
          description: "Number of order objects per page"
          required: false
      responses:
        200:
          description: "Successful Response"
          schema:
              $ref: "#/definitions/Order"
  /orders/{order_id}:
    get:
      tags: [
        "Orders"
      ]
      summary: "Get Order Endpoint"
      description: "Get one order's data by ID"
      parameters:
        - name: order_id
          in: path
          description: "ID of order to fetch"
      responses:
        200:
          description: "Successful Response"
          schema:
            $ref: "#/definitions/Order"
...
definitions: 
...
  Customer:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Customer's full name"
        example: "James Carlson"
      email:
        type: "string"
        description: "Customer's email address"
        example: "jamesmbcarlson@gmail.com"
      phone:
        type: "string"
        description: "Customer's phone number"
        example: "(719)555-8556"
      username:
        type: "string"
        description: "Customer's username"
        example: "jimjamsupreme"
  CustomerResponse:
    allOf:
      - $ref: '#/definitions/Customer'
      - type:
        properties:
          id:
            type: "integer"
            description: "Customer's unique ID number"
            example: "1"
            readonly: true
  CustomerRequest:
    allOf:
      - $ref: "#/definitions/Customer"
      - type:
        properties:
          password:
            type: "string"
            description: "Customer's password"
            example: "Password123!"
...
  Order:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Order's unique ID number"
        example: "42"
      customer_id:
        type: "integer"
        description: "Associated customer's unique ID number"
        example: "1"
      products:
        type: "array"
        description: "List of ordered products"
        items:
          type: "object"
          properties:
            id:
              type: "integer"
              description: "Ordered product's unique ID number"
              example: "7"
...